Department delete related bug fixed.

+++ src/services/department.service.ts
@@ -7,6 +7,7 @@ import validate from "../validators/validation";
 import { buildQuery } from "./util.service";
 import { IQuery } from "../interfaces/query.interface";
 import { IListResponse } from "../interfaces/response.interface";
+import User from "../models/user.model";
 
 const getDepartments = async (req: Request): Promise<IListResponse> => {
   const { query, queryParams } = buildQuery(QueryBuilderKeys.DEPARTMENT_LIST, req, { sort: "name", sortBy: SortBy.ASC } as IQuery);
@@ -61,6 +62,11 @@ const deleteDepartment = async (id: string): Promise<any> => {
   const department = await getSingleDepartment(id);
   if (!department) {
     throw new AppError(HttpStatus.BAD_REQUEST, AppMessages.DEPARTMENT_NOT_EXIST);
+  }
+
+  const userWithDepartment = await User.findOne({ department: id });
+  if (userWithDepartment) {
+    throw new AppError(HttpStatus.BAD_REQUEST, AppMessages.USER_WITH_DEPARTMENT_EXIST);
   }
 
   await Department.deleteOne({ _id: id });
